requirements.txt
typer>=0.12.0
rich>=13.0.0
loguru>=0.7.0
pydantic>=2.0.0
ruamel.yaml>=0.18.0
tqdm>=4.0.0
# PDF Core
pdfplumber>=0.11.0
pikepdf>=9.0.0
pymupdf>=1.24.0
ocrmypdf>=16.0.0
pytesseract>=0.3.10
# Tables & Data
camelot-py[cv]>=0.11.0
opencv-python>=4.10.0
# Utilities
regex>=2024.0.0
unidecode>=1.3.0
beautifulsoup4>=4.12.0
markdown-it-py>=3.0.0
pdf2image>=1.17.0
